studyConfig-improvements
1. consents component now has all the consesnts in it
