+++ src/pages/home/StudyConfig.vue
@@ -32,45 +32,13 @@
             </q-item-section>
           </q-item>
         </q-list>
-        <q-list v-if="studyDescription.consent.extraItems">
-          <q-item v-for="(extraItem, extraIndex) in studyDescription.consent.extraItems" :key="extraIndex">
-            <q-item-section>
-              <q-item-label>{{extraItem.description[$i18n.locale]}}</q-item-label>
-            </q-item-section>
-            <q-item-section avatar>
-              <q-checkbox v-model="consentedExtraItems[extraIndex]" :disabled="!extraItem.optional"/>
-            </q-item-section>
-          </q-item>
-        </q-list>
-        <q-list v-for="(taskItem, taskIndex) in studyDescription.consent.taskItems" :key="taskIndex">
-          <q-item>
-          <q-item-section >
-            <q-item-label class="q-my-md">{{taskItem.description[$i18n.locale]}}</q-item-label>
-              <q-item-label v-if="taskType[taskIndex] !== 'form' && !permissionsGiven[taskIndex] || taskType[taskIndex] !== 'form' && !consentedTaskItems[taskIndex]">
-                <div class="q-mt-sm text-secondary">
-                  {{ $t('studies.consent.OSpermission') }}
-                  {{ $t('studies.consent.giveOSPermission') }}
-                </div>
-              </q-item-label>
-          </q-item-section>
-            <q-item-section avatar>
-              <q-checkbox @click.native="(consentedTaskItems[taskIndex] || !permissionsGiven[taskIndex]) ? requestDQPermission(taskIndex) : null" v-model="consentedTaskItems[taskIndex]"/>
-            </q-item-section>
-          </q-item>
-        </q-list>
-        <q-separator/>
-        <q-list>
-          <q-item>
-            <q-item-section>
-              <q-item-label>{{ $t('studies.consent.remindersConsent') }}</q-item-label>
-              <q-item-label caption>{{ $t('studies.consent.remindersOSPermission') }}</q-item-label>
-              <q-btn class="q-mt-lg" :label="$t('studies.consent.giveRemindersOSPermission')" :disabled="!reminders || remindersPermissionGiven" @click="requestNotificationsPermission()"></q-btn>
-            </q-item-section>
-            <q-item-section avatar>
-              <q-checkbox v-model="reminders"/>
-            </q-item-section>
-          </q-item>
-        </q-list>
+      <consents
+        :studyDescription="studyDescription" :taskType="taskType"
+        :consentedExtraItems="consentedExtraItems" :consentedTaskItems="consentedTaskItems"
+        :remindersPermissionGiven="remindersPermissionGiven" :remindersPermissionNeeded="remindersPermissionNeeded"
+        @checkboxReminders="requestNotificationsPermission"
+      />
+
         <div class="q-my-md row justify-center">
           <q-btn :label="$t('studies.consent.updateConsent')" color="primary" :disabled="!canUpdate" @click="updateConsent()"></q-btn>
         </div>
@@ -83,48 +51,48 @@
 </template>
 
 <script>
+import Consents from 'components/Consents.vue'
 import StudyInfo from 'components/StudyInfo'
 import userinfo from 'modules/userinfo'
 import DB from 'modules/db'
 import API from 'modules/API'
-import healthStore from 'modules/healthstore'
 import notifications from 'modules/notifications'
 
 export default {
   name: 'StudyConfigPage',
   props: ['studyDescription'],
-  components: { StudyInfo },
+  components: { StudyInfo, Consents },
   data () {
     return {
       consentedExtraItems: [],
       consentedTaskItems: [],
-      permissionsGiven: [],
       studyParticipation: {},
-      reminders: false,
-      remindersPermissionNeeded: true,
-      remindersPermissionGiven: false,
+      remindersPermissionNeeded: Boolean,
+      remindersPermissionGiven: Boolean,
       tab: 'info'
     }
   },
   async created () {
     this.studyParticipation = await DB.getStudyParticipation(this.studyDescription._key)
-    // if it's wothdrawn or completed we don't need to show the following
+    // if it's withdrawn or completed we don't need to show the following
     if (this.studyParticipation.currentStatus !== 'withdrawn' && this.studyParticipation.currentStatus !== 'completed') {
-      this.reminders = this.studyParticipation.reminders
-      if (this.reminders) {
+      if (this.studyParticipation.reminders) {
         this.remindersPermissionNeeded = false
         this.remindersPermissionGiven = true
       } else {
-        let hasPermissionsAlready = await notifications.hasPermission()
-        this.remindersPermissionNeeded = !hasPermissionsAlready
-        this.remindersPermissionGiven = hasPermissionsAlready
+        let hasPermissionsAlready = await notifications.hasPermission() // true : false
+        this.remindersPermissionNeeded = !hasPermissionsAlready // false : true
+        this.remindersPermissionGiven = hasPermissionsAlready // true : false
       }
-      for (let i = 0; i < this.studyDescription.consent.extraItems.length; i++) {
-        this.consentedExtraItems.push(this.studyParticipation.extraItemsConsent[i].consented)
+      if (this.studyDescription.consent.extraItems) {
+        for (let i = 0; i < this.studyDescription.consent.extraItems.length; i++) {
+          this.consentedExtraItems.push(this.studyParticipation.extraItemsConsent[i].consented)
+        }
       }
-      for (let i = 0; i < this.studyDescription.consent.taskItems.length; i++) {
-        this.consentedTaskItems.push(this.studyParticipation.taskItemsConsent[i].consented)
-        this.permissionsGiven.push(this.studyParticipation.taskItemsConsent[i].consented)
+      if (this.studyDescription.consent.taskItems) {
+        for (let i = 0; i < this.studyDescription.consent.taskItems.length; i++) {
+          this.consentedTaskItems.push(this.studyParticipation.taskItemsConsent[i].consented)
+        }
       }
     }
   },
@@ -135,32 +103,33 @@ export default {
       })
     },
     canUpdate () {
-      if (this.reminders !== this.studyParticipation.reminders) {
-        if (this.reminders && !this.remindersPermissionGiven) return false
-        return true
-      }
-      if (!this.studyParticipation || !this.studyParticipation.extraItemsConsent) return false
-      for (let i = 0; i < this.studyParticipation.extraItemsConsent.length; i++) {
-        if (this.consentedExtraItems[i] !== this.studyParticipation.extraItemsConsent[i].consented) {
-          return true
+      if (this.studyParticipation.taskItemsConsent) {
+        for (let i = 0; i < this.studyParticipation.taskItemsConsent.length; i++) {
+          if (this.consentedTaskItems[i] !== this.studyParticipation.taskItemsConsent[i].consented) {
+            return true
+          }
         }
       }
-      for (let i = 0; i < this.studyParticipation.taskItemsConsent.length; i++) {
-        if (this.consentedTaskItems[i] !== this.studyParticipation.taskItemsConsent[i].consented && this.permissionsGiven[i]) {
-          return true
+      if (this.studyParticipation.extraItemsConsent) {
+        for (let i = 0; i < this.studyParticipation.extraItemsConsent.length; i++) {
+          if (this.consentedExtraItems[i] !== this.studyParticipation.extraItemsConsent[i].consented) {
+            return true
+          }
         }
       }
-      return false
+      return this.remindersPermissionGiven !== this.studyParticipation.reminders
     }
   },
   methods: {
     async updateConsent () {
-      this.studyParticipation.reminders = this.reminders
-      for (let i = 0; i < this.studyDescription.consent.extraItems.length; i++) {
-        this.studyParticipation.extraItemsConsent[i].consented = this.consentedExtraItems[i]
-      }
+      this.studyParticipation.reminders = this.remindersPermissionGiven
       for (let i = 0; i < this.studyDescription.consent.taskItems.length; i++) {
         this.studyParticipation.taskItemsConsent[i].consented = this.consentedTaskItems[i]
+      }
+      if (this.studyParticipation.extraItemsConsent) {
+        for (let i = 0; i < this.studyDescription.consent.extraItems.length; i++) {
+          this.studyParticipation.extraItemsConsent[i].consented = this.consentedExtraItems[i]
+        }
       }
       // call the API
       await API.updateStudyStatus(userinfo.user._key, this.studyParticipation.studyKey, this.studyParticipation)
@@ -171,43 +140,25 @@ export default {
         icon: 'check'
       })
     },
-    async requestDQPermission (taskIndex) {
-      let taskId = this.studyDescription.consent.taskItems[taskIndex].taskId
-      let taskdescr = this.studyDescription.tasks.find(t => t.id === taskId)
-      if (taskdescr && taskdescr.dataType) {
+    async requestNotificationsPermission (hasRemindersPermission) {
+      if (hasRemindersPermission) {
+        this.remindersPermissionGiven = false
+      } else {
         try {
-          await healthStore.requestAuthorization([taskdescr.dataType])
-          this.$set(this.permissionsGiven, taskIndex, true)
+          this.remindersPermissionGiven = await notifications.requestPermission()
           this.$q.notify({
             color: 'positive',
             message: this.$i18n.t('studies.consent.OSPermissionGiven'),
             icon: 'check'
           })
         } catch (error) {
-          console.error('Cannot get authorisation for health', error)
+          console.error('Cannot get authorisation for sending reminders', error)
           this.$q.notify({
             color: 'negative',
             message: this.$i18n.t('studies.consent.OSPermissionNotGiven') + ': ' + error.message,
             icon: 'report_problem'
           })
         }
-      }
-    },
-    async requestNotificationsPermission () {
-      try {
-        this.remindersPermissionGiven = await notifications.requestPermission()
-        this.$q.notify({
-          color: 'positive',
-          message: this.$i18n.t('studies.consent.OSPermissionGiven'),
-          icon: 'check'
-        })
-      } catch (error) {
-        console.error('Cannot get authorisation for sending reminders', error)
-        this.$q.notify({
-          color: 'negative',
-          message: this.$i18n.t('studies.consent.OSPermissionNotGiven') + ': ' + error.message,
-          icon: 'report_problem'
-        })
       }
     },
     async withdraw () {
